# Conflicts:
#	.github/dependabot.yml

+++ test/FluentMigrator.Tests/Integration/Processors/Firebird/TemporaryDatabase.cs
@@ -29,8 +29,19 @@ namespace FluentMigrator.Tests.Integration.Processors.Firebird
 
         public TemporaryDatabase(IntegrationTestOptions.DatabaseServerOptions connectionOptions, FirebirdLibraryProber prober)
         {
-            if (!connectionOptions.IsEnabled)
-                Assert.Ignore();
+            connectionOptions.IgnoreIfNotEnabled();
+
+            if (connectionOptions.ContainerEnabled)
+            {
+                _connectionString = new Lazy<string>(() =>
+                {
+                    var csb = new FbConnectionStringBuilder(connectionOptions.ConnectionString);
+
+                    return prober.CreateDb(csb);
+                });
+
+                return;
+            }
 
             DbFileName = Path.GetTempFileName();
             _connectionString = new Lazy<string>(() =>
@@ -53,7 +64,12 @@ namespace FluentMigrator.Tests.Integration.Processors.Firebird
         {
             try
             {
-                File.Delete(DbFileName);
+                if (!string.IsNullOrEmpty(DbFileName))
+                {
+                    File.Delete(DbFileName);
+                }
+
+                FbDatabase.DropDatabase(ConnectionString);
             }
             catch
             {
